+++ src/middlewares/unhandledExceptions.ts
@@ -1,10 +1,22 @@
-import * as express from 'express';
+import { ExceptionFilter, Catch, ArgumentsHost, HttpException } from '@nestjs/common';
+import { Request, Response } from 'express';
 import logger from '../common/logger'
 
-export default (err: express.Errback, _: express.Request, response: express.Response, next: express.NextFunction) => {
-  if (err) {
-    logger.logError("Something happend in service", err);
-    response.statusCode = 500;
+@Catch(HttpException)
+export class HttpExceptionFilter implements ExceptionFilter {
+  catch(exception: HttpException, host: ArgumentsHost) {
+    const ctx = host.switchToHttp();
+    const response = ctx.getResponse<Response>();
+    const request = ctx.getRequest<Request>();
+    const status = exception.getStatus();
+
+    logger.logError("Something happend in service", exception.message);
+    response
+      .status(500)
+      .json({
+        statusCode: status,
+        timestamp: new Date().toISOString(),
+        path: request.url,
+      });
   }
-  next(err);
 }
